user-service
user-service:
url: localhost:4001/:1/updateuser
method: post
body:
{
    "username": "asdfj",
    "password": "must be hashed before using api",
    "email": "email",
    "role": "role",
    "createdAt": "createdAt",
    "updatedAt": "updatedAt"
}

take note that this is somewhat dynamic. only pass in values you want to use to update the user. id is not passed in, i assume you never need to change that. id is being passed into url, not sure if we want to keep it this way. we should have a better view for this once front end is able to interact with this.